chore(backend): changed logger util func to different dir
